+++ src/app/metadata/metadataRegistry/datasetMetadataComponent.ts
@@ -1,5 +1,6 @@
 import { Component, EventEmitter, Input, Output, SimpleChanges } from "@angular/core";
 import { forkJoin, Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 import { ARTLiteral, ARTURIResource } from 'src/app/models/ARTResources';
 import { MdrVoc } from 'src/app/models/Vocabulary';
 import { ResourceUtils } from 'src/app/utils/ResourceUtils';
@@ -78,39 +79,52 @@ export class DatasetMetadataComponent {
         // normalize limitation
         this.sparqlLimitations = false;
         if (this.dataset.sparqlEndpoint.limitations != null) {
-            this.sparqlLimitations = this.dataset.sparqlEndpoint.limitations.indexOf(MdrVoc.noAggregation.toNT()) != -1;
+            this.sparqlLimitations = this.dataset.sparqlEndpoint.limitations.indexOf(MdrVoc.noAggregation.getURI()) != -1;
         }
     }
 
     editTitles() {
-        let localizedMap: LocalizedMap = {};
-        this.dataset.titles.forEach(t => { localizedMap[t.getLang()] = t.getValue(); });
+        let localizedMap: LocalizedMap = new Map();
+        this.dataset.titles.forEach(t => { localizedMap.set(t.getLang(), t.getValue()); });
         this.sharedModals.localizedEditor({ key: "COMMONS.TITLE" }, localizedMap).then(
             (newLocalizedMap: LocalizedMap) => {
                 let toUpdate: ARTLiteral[] = [];
-                for (let lang in newLocalizedMap) {
+                newLocalizedMap.forEach((title, lang) => {
                     let old = this.dataset.titles.find(t => t.getLang() == lang);
-                    if (old == null || old.getValue() != newLocalizedMap[lang]) {
-                        toUpdate.push(new ARTLiteral(newLocalizedMap[lang], null, lang));
+                    if (old == null || old.getValue() != title) {
+                        toUpdate.push(new ARTLiteral(title, null, lang));
                     }
-                }
+                });
                 let toDelete: ARTLiteral[] = [];
                 this.dataset.titles.forEach(t => {
-                    if (newLocalizedMap[t.getLang()] == null) {
+                    if (newLocalizedMap.get(t.getLang()) == null) {
                         toDelete.push(t);
                     }
                 });
                 if (toUpdate.length > 0) {
                     let updateFn: Observable<void>[] = toUpdate.map(t => {
-                        return this.metadataRegistryService.setTitle(this.dataset.identity, t);
+                        return this.metadataRegistryService.setTitle(this.dataset.identity, t).pipe(
+                            map(() => {
+                                let updatedIdx = this.dataset.titles.findIndex(title => title.getLang() == t.getLang());
+                                if (updatedIdx == -1) {
+                                    this.dataset.titles.push(t); //new
+                                } else {
+                                    this.dataset.titles[updatedIdx] = t; //updated
+                                }
+                            })
+                        );
                     });
                     forkJoin(updateFn).subscribe();
                 }
                 if (toDelete.length > 0) {
-                    // let removeFn: Observable<void>[] = toDelete.map(t => {
-                    //     return this.metadataRegistryService.removeTitle(this.dataset.identity, t);
-                    // });
-                    // forkJoin(removeFn).subscribe();
+                    let removeFn: Observable<void>[] = toDelete.map(t => {
+                        return this.metadataRegistryService.deleteTitle(this.dataset.identity, t).pipe(
+                            map(() => {
+                                this.dataset.titles.splice(this.dataset.titles.indexOf(t), 1);
+                            })
+                        );
+                    });
+                    forkJoin(removeFn).subscribe();
                 }
             },
             () => { }
@@ -118,35 +132,47 @@ export class DatasetMetadataComponent {
     }
 
     editDescriptions() {
-        let localizedMap: LocalizedMap = {};
-        this.dataset.descriptions.forEach(d => { localizedMap[d.getLang()] = d.getValue(); });
+        let localizedMap: LocalizedMap = new Map();
+        this.dataset.descriptions.forEach(d => { localizedMap.set(d.getLang(), d.getValue()); });
         this.sharedModals.localizedEditor({ key: "COMMONS.DESCRIPTION" }, localizedMap, true).then(
             (newLocalizedMap: LocalizedMap) => {
                 let toUpdate: ARTLiteral[] = [];
-                for (let lang in newLocalizedMap) {
+                newLocalizedMap.forEach((descr, lang) => {
                     let old = this.dataset.descriptions.find(d => d.getLang() == lang);
-                    if (old == null || old.getValue() != newLocalizedMap[lang]) {
-                        toUpdate.push(new ARTLiteral(newLocalizedMap[lang], null, lang));
+                    if (old == null || old.getValue() != descr) {
+                        toUpdate.push(new ARTLiteral(descr, null, lang));
                     }
-                }
+                });
                 let toDelete: ARTLiteral[] = [];
                 this.dataset.descriptions.forEach(d => {
-                    if (newLocalizedMap[d.getLang()] == null) {
+                    if (newLocalizedMap.get(d.getLang()) == null) {
                         toDelete.push(d);
                     }
                 });
-
                 if (toUpdate.length > 0) {
-                    let updateFn: Observable<void>[] = toUpdate.map(t => {
-                        return this.metadataRegistryService.setDescription(this.dataset.identity, t);
+                    let updateFn: Observable<void>[] = toUpdate.map(d => {
+                        return this.metadataRegistryService.setDescription(this.dataset.identity, d).pipe(
+                            map(() => {
+                                let updatedIdx = this.dataset.descriptions.findIndex(descr => descr.getLang() == d.getLang());
+                                if (updatedIdx == -1) {
+                                    this.dataset.descriptions.push(d); //new
+                                } else {
+                                    this.dataset.descriptions[updatedIdx] = d; //updated
+                                }
+                            })
+                        );
                     });
                     forkJoin(updateFn).subscribe();
                 }
                 if (toDelete.length > 0) {
-                    // let removeFn: Observable<void>[] = toDelete.map(t => {
-                    //     return this.metadataRegistryService.removeDescription(this.dataset.identity, t);
-                    // });
-                    // forkJoin(removeFn).subscribe();
+                    let removeFn: Observable<void>[] = toDelete.map(d => {
+                        return this.metadataRegistryService.deleteDescription(this.dataset.identity, d).pipe(
+                            map(() => {
+                                this.dataset.descriptions.splice(this.dataset.descriptions.indexOf(d), 1);
+                            })
+                        );
+                    });
+                    forkJoin(removeFn).subscribe();
                 }
             },
             () => { }
@@ -229,6 +255,13 @@ export class DatasetMetadataComponent {
         }
         this.metadataRegistryService.setSPARQLEndpoint(this.dataset.identity, sparqlEndpoint).subscribe(
             () => {
+                if (this.dataset.sparqlEndpoint) {
+                    this.dataset.sparqlEndpoint.id = sparqlEndpoint.getURI();
+                } else {
+                    this.dataset.sparqlEndpoint = {
+                        id: sparqlEndpoint.getURI(),
+                    };
+                }
                 this.initDatasetMetadata();
                 this.update.emit();
             }
@@ -248,6 +281,7 @@ export class DatasetMetadataComponent {
         }
         this.metadataRegistryService.setDereferenciability(this.dataset.identity, dereferenciablePar).subscribe(
             () => {
+                this.dataset.dereferenciationSystem = newValue;
                 this.initDatasetMetadata();
                 this.update.emit();
             }
@@ -258,6 +292,7 @@ export class DatasetMetadataComponent {
         if (this.sparqlLimitations) {
             this.metadataRegistryService.setSPARQLEndpointLimitation(new ARTURIResource(this.dataset.sparqlEndpoint.id), MdrVoc.noAggregation).subscribe(
                 () => {
+                    this.dataset.sparqlEndpoint.limitations = [MdrVoc.noAggregation.getURI()];
                     this.initDatasetMetadata();
                     this.update.emit();
                 }
@@ -265,122 +300,12 @@ export class DatasetMetadataComponent {
         } else {
             this.metadataRegistryService.removeSPARQLEndpointLimitation(new ARTURIResource(this.dataset.sparqlEndpoint.id), MdrVoc.noAggregation).subscribe(
                 () => {
+                    this.dataset.sparqlEndpoint.limitations.splice(this.dataset.sparqlEndpoint.limitations.indexOf(MdrVoc.noAggregation.getURI()), 1);
                     this.initDatasetMetadata();
                     this.update.emit();
                 }
             );
         }
     }
-
-
-
-
-    //////
-
-    // @Input() dataset: DatasetMetadata;
-    // @Input() version: boolean = false; //tells if the DatasetMetadata represents a version. This deteremines whether to show/hide the "Version Info"
-    // @Output() update = new EventEmitter();
-
-    // private dereferUnknown: string = "Unknown";
-    // private dereferYes: string = "Yes";
-    // private dereferNo: string = "No";
-
-    // dereferenciationValues: string[] = [this.dereferUnknown, this.dereferYes, this.dereferNo];
-    // dereferenciationNormalized: string;
-
-    // sparqlLimitations: boolean;
-
-    // constructor(private metadataRegistryService: MetadataRegistryServices, private basicModals: BasicModalServices) { }
-
-    // ngOnChanges(changes: SimpleChanges) {
-    //     if (changes['dataset'] && changes['dataset'].currentValue) {
-    //         //normalize dereferenciation
-    //         if (this.dataset.dereferenciationSystem == null) {
-    //             this.dereferenciationNormalized = this.dereferUnknown;
-    //         } else if (this.dataset.dereferenciationSystem == MdrVoc.standardDereferenciation.getURI()) {
-    //             this.dereferenciationNormalized = this.dereferYes;
-    //         } else if (this.dataset.dereferenciationSystem == MdrVoc.noDereferenciation.getURI()) {
-    //             this.dereferenciationNormalized = this.dereferNo;
-    //         } else {
-    //             this.dereferenciationValues.push(this.dataset.dereferenciationSystem);
-    //             this.dereferenciationNormalized = this.dataset.dereferenciationSystem;
-    //         }
-    //         //normalize limitation
-    //         this.sparqlLimitations = false;
-    //         if (this.dataset.sparqlEndpointMetadata.limitations != null) {
-    //             this.sparqlLimitations = this.dataset.sparqlEndpointMetadata.limitations.indexOf(MdrVoc.noAggregation.toNT()) != -1;
-    //         }
-    //     }
-    // }
-
-    // updateTitle(newValue: string) {
-    //     let title: string = null;
-    //     if (newValue != null && newValue.trim() != "") {
-    //         title = newValue;
-    //     }
-    //     this.metadataRegistryService.setTitle(new ARTURIResource(this.dataset.identity), title).subscribe(
-    //         stResp => {
-    //             this.update.emit();
-    //         }
-    //     );
-    // }
-
-    // updateSparqlEndpoint(newValue: string) {
-    //     let sparqlEndpoint: ARTURIResource;
-    //     if (newValue != null && newValue.trim() != "") {
-    //         if (ResourceUtils.testIRI(newValue)) {
-    //             sparqlEndpoint = new ARTURIResource(newValue);
-    //         } else { //invalid IRI
-    //             this.basicModals.alert({ key: "STATUS.INVALID_VALUE" }, { key: "MESSAGES.INVALID_IRI", params: { iri: newValue } }, ModalType.warning);
-    //             //restore old id
-    //             let backupId: string = this.dataset.sparqlEndpointMetadata.id;
-    //             this.dataset.sparqlEndpointMetadata.id = null + "new";
-    //             setTimeout(() => {
-    //                 this.dataset.sparqlEndpointMetadata.id = backupId;
-    //             });
-    //             return;
-    //         }
-    //     }
-    //     this.metadataRegistryService.setSPARQLEndpoint(new ARTURIResource(this.dataset.identity), sparqlEndpoint).subscribe(
-    //         stResp => {
-    //             this.update.emit();
-    //         }
-    //     );
-    // }
-
-    // updateDerefSystem(newValue: string) {
-    //     let dereferenciablePar: boolean;
-    //     if (newValue == this.dereferUnknown) {
-    //         dereferenciablePar = null;
-    //     } else if (newValue == this.dereferYes) {
-    //         dereferenciablePar = true;
-    //     } else if (newValue == this.dereferNo) {
-    //         dereferenciablePar = false;
-    //     } else { //custom choice, available only if it was already the dereferenciationSystem, so it wasn't canged
-    //         return;
-    //     }
-    //     this.metadataRegistryService.setDereferenciability(new ARTURIResource(this.dataset.identity), dereferenciablePar).subscribe(
-    //         () => {
-    //             this.update.emit();
-    //         }
-    //     );
-    // }
-
-    // updateSparqlLimitations() {
-    //     if (this.sparqlLimitations) {
-    //         this.metadataRegistryService.setSPARQLEndpointLimitation(new ARTURIResource(this.dataset.sparqlEndpointMetadata.id), MdrVoc.noAggregation).subscribe(
-    //                 stResp => {
-    //                     this.update.emit();
-    //                 }
-    //             );
-    //     } else {
-    //         this.metadataRegistryService.removeSPARQLEndpointLimitation(new ARTURIResource(this.dataset.sparqlEndpointMetadata.id), MdrVoc.noAggregation).subscribe(
-    //                 stResp => {
-    //                     this.update.emit();
-    //                 }
-    //             );
-    //     }
-
-    // }
 
 }