+++ ChangeLog.md
@@ -6,7 +6,9 @@ https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
 
 # 11.0.1 (dd-mm-2022)
   * Added support for EXT project universal access level
+  * Improved filters in users manager page
   * Restored image display and foreign URIs resolution in ResourceView (erroneously disabled in v11.0.0)
+  * Fixed several bugs in Metadata Registry that affected multiple features in VB
 
 # 11.0.0 (12-05-2022)
   * Added SuperUser system role: super users create projects where it is automatically set as Project Manager